+++ src/backends/sentry_backend_inproc.c
@@ -1104,7 +1104,7 @@ process_ucontext_deferred(const sentry_ucontext_t *uctx,
             sentry_envelope_t *envelope = sentry__prepare_event(options, event,
                 NULL, !options->on_crash_func && !skip_hooks, NULL);
             if (envelope) {
-                event_id = sentry__envelope_get_event_id(envelope);
+                event_id = sentry_envelope_get_event_id(envelope);
             }
             sentry_session_t *session = sentry__end_current_session_with_status(
                 SENTRY_SESSION_STATUS_CRASHED);