+++ src/sentry_envelope.c
@@ -296,7 +296,7 @@ sentry__envelope_from_path(const sentry_path_t *path)
 }
 
 sentry_uuid_t
-sentry__envelope_get_event_id(const sentry_envelope_t *envelope)
+sentry_envelope_get_event_id(const sentry_envelope_t *envelope)
 {
     if (!envelope) {
         return sentry_uuid_nil();
@@ -1275,7 +1275,7 @@ sentry__envelope_write_to_cache(
         return 1;
     }
 
-    sentry_uuid_t event_id = sentry__envelope_get_event_id(envelope);
+    sentry_uuid_t event_id = sentry_envelope_get_event_id(envelope);
     if (sentry_uuid_is_nil(&event_id)) {
         event_id = sentry_uuid_new_v4();
     }